fix(invoices): replace float with Decimal in stock deduction and serializer

+++ my_iveco_crm/invoices/serializers.py
@@ -16,7 +16,7 @@ class InvoiceItemSerializer(serializers.ModelSerializer):
         read_only_fields = ['id', 'product_name', 'product_sku', 'total']
 
     def get_total(self, obj):
-        return float(obj.quantity * obj.unit_price)
+        return obj.quantity * obj.unit_price
 
 
 class InvoiceSerializer(serializers.ModelSerializer):
